Fix phantom-edges
Skip adding "potential" phantom edges to the graph—as a result, real phantom
vertices will not get added to the graph from either of its container subgraphs.
But false-phantom vertices will have a real edge missing at any of its sides.
This is fixed during the contraction, by tracking whether a vertex is seen in
two edges or not. If not, add the false-phantom edge to the graph at this stage.